Add files via upload
NFT COLLECTION (Metacrafter)

Create a variable that can hold a number of NFT's. What type of variable might this be?
Create an object inside your mintNFT function that will hold the metadata for your NFTs. The metadata values will be passed to the function as parameters. When the NFT is ready, you will store it in the variable you created in step 2
Your listNFTs() function will print all of your NFTs metadata to the console (i.e. console.log("Name: " + someNFT.name))
For good measure, getTotalSupply() should return the number of NFT's you have created